requirements.txt
pdfplumber==0.11.2
python-docx==1.1.0
python-pptx==0.6.23
bs4==0.0.2
beautifulsoup4==4.12.3
pandas==2.2.2
openpyxl==3.1.2
lxml==5.2.2
python-dotenv==1.0.1
pytest==8.3.2
tqdm==4.66.4
numpy==2.1.2
streamlit==1.42.0
faiss-cpu==1.7.3    
 
